refactor tbank utils and types
