+++ backend/apps/tbank/managers/payment.py
@@ -4,7 +4,8 @@ import logging
 from django.conf import settings
 from django.db.models import Manager
 
-from apps.tbank.classes.TBank import TBank, OperationInitiatorType, ReceiptFFD12, ReceiptFFD105
+from apps.tbank.classes.TBank import TBank
+from apps.tbank.types import OperationInitiatorType, ReceiptFFD12, ReceiptFFD105
 from apps.tbank.exceptions.base import TBankException
 
 log = logging.getLogger('tbank')
